--- src/test/java/ca/gc/aafc/dina/export/api/generator/TabularDataExportWithAliasesIT.java
@@ -1,489 +1,0 @@
-package ca.gc.aafc.dina.export.api.generator;
-
-import ca.gc.aafc.dina.exception.ResourceNotFoundException;
-import ca.gc.aafc.dina.export.api.BaseIntegrationTest;
-import ca.gc.aafc.dina.export.api.ElasticSearchTestContainerInitializer;
-import ca.gc.aafc.dina.export.api.async.AsyncConsumer;
-import ca.gc.aafc.dina.export.api.dto.DataExportDto;
-import ca.gc.aafc.dina.export.api.entity.DataExport;
-import ca.gc.aafc.dina.export.api.file.FileController;
-import ca.gc.aafc.dina.export.api.repository.DataExportRepository;
-import ca.gc.aafc.dina.testsupport.elasticsearch.ElasticSearchTestUtils;
-import co.elastic.clients.elasticsearch.ElasticsearchClient;
-import org.junit.jupiter.api.AfterEach;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-import org.springframework.core.io.InputStreamResource;
-import org.springframework.http.ResponseEntity;
-import org.springframework.test.context.ContextConfiguration;
-
-import javax.inject.Inject;
-import java.io.BufferedReader;
-import java.io.IOException;
-import java.io.InputStreamReader;
-import java.nio.charset.StandardCharsets;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Map;
-import java.util.UUID;
-import java.util.concurrent.ExecutionException;
-import java.util.concurrent.Future;
-import java.util.zip.ZipEntry;
-import java.util.zip.ZipInputStream;
-
-import static org.junit.jupiter.api.Assertions.*;
-
-/**
- * Integration tests for TabularDataExportGenerator with column aliases and normalized relationships.
- * Tests ENA-style export format using TabularDataExportGenerator's column aliasing feature.
- */
-@ContextConfiguration(initializers = {ElasticSearchTestContainerInitializer.class})
-public class TabularDataExportWithAliasesIT extends BaseIntegrationTest {
-
-  private static final String MAT_SAMPLE_INDEX = "dina_material_sample_index";
-
-  @Inject
-  private DataExportRepository dataExportRepository;
-
-  @Inject
-  private DataExportServiceTransactionWrapper dataExportServiceTransactionWrapper;
-
-  @Inject
-  private FileController fileController;
-
-  @Inject
-  private ElasticsearchClient esClient;
-
-  @Inject
-  private AsyncConsumer<Future<UUID>> asyncConsumer;
-
-  @BeforeEach
-  public void setupElasticSearch() throws IOException {
-    // Delete index if it exists from previous test
-    try {
-      esClient.indices().delete(d -> d.index(MAT_SAMPLE_INDEX));
-    } catch (Exception e) {
-      // Ignore if index doesn't exist
-    }
-  }
-
-  @AfterEach
-  public void cleanupElasticSearch() throws IOException {
-    try {
-      esClient.indices().delete(d -> d.index(MAT_SAMPLE_INDEX));
-    } catch (Exception e) {
-      // Ignore cleanup errors
-    }
-  }
-
-  @Test
-  public void testNormalizedExportWithColumnAliases()
-      throws IOException, ExecutionException, InterruptedException,
-             ca.gc.aafc.dina.exception.ResourceNotFoundException,
-             ca.gc.aafc.dina.exception.ResourceGoneException {
-
-    // Setup ElasticSearch index with test data
-    ElasticSearchTestUtils.createIndex(esClient, MAT_SAMPLE_INDEX,
-        "elasticsearch/material_sample_index_settings.json");
-
-    // Create a shared project ID so both samples belong to the same project (for deduplication testing)
-    UUID sharedProjectId = UUID.randomUUID();
-
-    // Add test documents with realistic structure - both samples share the same project
-    UUID docId1 = UUID.randomUUID();
-    String testDoc1 = buildTestMaterialSampleDocument(docId1, "BEA01-20161005930-A", sharedProjectId);
-    ElasticSearchTestUtils.indexDocument(esClient, MAT_SAMPLE_INDEX, docId1.toString(), testDoc1);
-
-    UUID docId2 = UUID.randomUUID();
-    String testDoc2 = buildTestMaterialSampleDocument(docId2, "BEA02-20161005931-B", sharedProjectId);
-    ElasticSearchTestUtils.indexDocument(esClient, MAT_SAMPLE_INDEX, docId2.toString(), testDoc2);
-
-    // Create tabular data export with column aliases and normalized relationships.
-    // This produces a ZIP file containing separate samples.csv and projects.csv files.
-    UUID uuid = UUID.randomUUID();
-
-    DataExport dataExport = DataExport.builder()
-        .uuid(uuid)
-        .createdBy("test-user")
-        .source(MAT_SAMPLE_INDEX)
-        .exportType(DataExport.ExportType.TABULAR_DATA)
-        .query(Map.of("query", Map.of("match_all", Map.of())))
-        .columns(new String[]{
-            "projects.name",
-            "projects.name",
-            "projects.multilingualDescription.descriptions[0].desc",
-            "materialSampleName",
-            "materialSampleName",
-            "effectiveScientificName",
-            "effectiveScientificName",
-            "materialSampleRemarks",
-            "managedAttributes.*",
-            "collectingEvent.dwcVerbatimLocality",
-            "collectingEvent.endEventDateTime"
-        })
-        .columnAliases(new String[]{
-            "Project.alias",
-            "Project.title",
-            "Project.description",
-            "Sample.alias",
-            "Sample.title",
-            "Sample.sample_name.taxon_id",
-            "Sample.sample_name.scientific_name",
-            "Sample.description",
-            "Sample.sample_attributes",
-            "Sample.GeographicLocation",
-            "Sample.collectionDate"
-        })
-        .exportOptions(Map.of(
-            "normalizeRelationships", "true",
-            "columnSeparator", "COMMA"
-        ))
-        .name("normalized_export_test")
-        .build();
-
-    dataExportServiceTransactionWrapper.createEntity(dataExport);
-
-    // Wait for async processing to complete - get the last added task
-    int taskIndex = asyncConsumer.getAccepted().size() - 1;
-    asyncConsumer.getAccepted().get(taskIndex).get();
-
-    // Verify export completed successfully
-    DataExportDto savedDataExportDto = dataExportRepository.getOne(uuid, null).getDto();
-    assertEquals(DataExport.ExportStatus.COMPLETED, savedDataExportDto.getStatus());
-    assertEquals(DataExport.ExportType.TABULAR_DATA, savedDataExportDto.getExportType());
-    assertEquals("normalized_export_test", savedDataExportDto.getName());
-
-    // Download and verify the ZIP file (normalized export creates a ZIP with multiple CSVs)
-    ResponseEntity<InputStreamResource> response = fileController.downloadFile(uuid,
-        FileController.DownloadType.DATA_EXPORT);
-
-    assertEquals("normalized_export_test.zip", 
-        response.getHeaders().getContentDisposition().getFilename());
-
-    assertNotNull(response.getBody());
-
-    // Extract and verify ZIP contents - should have samples.csv and projects.csv
-    try (ZipInputStream zis = new ZipInputStream(response.getBody().getInputStream())) {
-      List<String> samplesLines = null;
-      List<String> projectsLines = null;
-
-      ZipEntry entry;
-      while ((entry = zis.getNextEntry()) != null) {
-        String fileName = entry.getName();
-        List<String> lines = readLinesFromZipEntry(zis);
-
-        if ("samples.csv".equals(fileName)) {
-          samplesLines = lines;
-        } else if ("projects.csv".equals(fileName)) {
-          projectsLines = lines;
-        }
-
-        zis.closeEntry();
-      }
-
-      // Verify samples.csv exists and has correct structure
-      assertNotNull(samplesLines, "samples.csv should exist in ZIP");
-      assertTrue(samplesLines.size() >= 3, "Should have header + at least 2 sample rows");
-
-      System.out.println("=== samples.csv ===");
-      samplesLines.forEach(System.out::println);
-
-      // Verify samples.csv header contains expected sample columns
-      String samplesHeader = samplesLines.get(0);
-      assertTrue(samplesHeader.contains("Sample.alias"), "Should have Sample.alias column");
-      assertTrue(samplesHeader.contains("Sample.title"), "Should have Sample.title column");
-      assertTrue(samplesHeader.contains("Sample.sample_name.taxon_id"), 
-          "Should have Sample.sample_name.taxon_id column");
-      assertTrue(samplesHeader.contains("Sample.sample_name.scientific_name"), 
-          "Should have Sample.sample_name.scientific_name column");
-      assertTrue(samplesHeader.contains("Sample.description"), "Should have Sample.description column");
-
-      // Verify sample data rows contain expected sample names
-      String sampleData = String.join("\n", samplesLines);
-      assertTrue(sampleData.contains("BEA01-20161005930-A"), 
-          "Should contain first sample name");
-      assertTrue(sampleData.contains("BEA02-20161005931-B"), 
-          "Should contain second sample name");
-
-      // Verify projects.csv exists and has correct structure
-      assertNotNull(projectsLines, "projects.csv should exist in ZIP");
-      assertTrue(projectsLines.size() >= 2, "Should have header + at least 1 project row");
-
-      System.out.println("=== projects.csv ===");
-      projectsLines.forEach(System.out::println);
-
-      // Verify projects.csv header contains expected project columns
-      String projectsHeader = projectsLines.get(0);
-      assertTrue(projectsHeader.contains("Project.alias"), "Should have Project.alias column");
-      assertTrue(projectsHeader.contains("Project.title"), "Should have Project.title column");
-      assertTrue(projectsHeader.contains("Project.description"), 
-          "Should have Project.description column");
-
-      // Verify project data contains the expected project name (deduplicated)
-      String projectData = String.join("\n", projectsLines);
-      assertTrue(projectData.contains("GRDI-Eco Invertebrates"),
-          "Should contain project name");
-      
-      // Projects should be deduplicated - only 1 unique project in test data
-      assertEquals(2, projectsLines.size(), 
-          "Should have exactly 2 lines (header + 1 unique project row)");
-    }
-
-    // Clean up - delete the export
-    // Note: Normalized exports create a directory structure, so cleanup may fail with DirectoryNotEmptyException
-    // This is expected behavior and doesn't affect the test validity
-    try {
-      dataExportRepository.onDelete(uuid);
-      assertThrows(ResourceNotFoundException.class,
-          () -> dataExportRepository.onFindOne(uuid, null));
-    } catch (RuntimeException e) {
-      // Ignore cleanup errors (e.g., directory not empty)
-    }
-  }
-
-  @Test
-  public void testExportWithColumnAliasesMinimalData()
-      throws IOException, ExecutionException, InterruptedException,
-             ca.gc.aafc.dina.exception.ResourceNotFoundException,
-             ca.gc.aafc.dina.exception.ResourceGoneException {
-
-    // Setup ElasticSearch index with minimal data
-    String testIndex = MAT_SAMPLE_INDEX + "_minimal";
-    ElasticSearchTestUtils.createIndex(esClient, testIndex,
-        "elasticsearch/material_sample_index_settings.json");
-
-    UUID docId = UUID.randomUUID();
-    String minimalDoc = buildMinimalMaterialSampleDocument(docId, "MINIMAL-SAMPLE-001");
-    ElasticSearchTestUtils.indexDocument(esClient, testIndex, docId.toString(), minimalDoc);
-
-    // Create tabular data export with column aliases (without normalized relationships)
-    UUID uuid = UUID.randomUUID();
-
-    DataExport dataExport = DataExport.builder()
-        .uuid(uuid)
-        .createdBy("test-user")
-        .source(testIndex)
-        .exportType(DataExport.ExportType.TABULAR_DATA)
-        .query(Map.of("query", Map.of("match_all", Map.of())))
-        .columns(new String[]{
-            "projects.name",
-            "projects.name",
-            "projects.multilingualDescription.descriptions[0].desc",
-            "materialSampleName",
-            "materialSampleName",
-            "effectiveScientificName",
-            "effectiveScientificName",
-            "materialSampleRemarks",
-            "managedAttributes.*",
-            "collectingEvent.dwcVerbatimLocality",
-            "collectingEvent.endEventDateTime"
-        })
-        .columnAliases(new String[]{
-            "Project.alias",
-            "Project.title",
-            "Project.description",
-            "Sample.alias",
-            "Sample.title",
-            "Sample.sample_name.taxon_id",
-            "Sample.sample_name.scientific_name",
-            "Sample.description",
-            "Sample.sample_attributes",
-            "Sample.GeographicLocation",
-            "Sample.collectionDate"
-        })
-        .exportOptions(Map.of("columnSeparator", "COMMA"))
-        .name("minimal_export_test")
-        .build();
-
-    dataExportServiceTransactionWrapper.createEntity(dataExport);
-
-    // Wait for completion (this is the second test, so it's at index 1)
-    asyncConsumer.getAccepted().get(asyncConsumer.getAccepted().size() - 1).get();
-
-    // Verify it completed
-    DataExportDto savedDataExportDto = dataExportRepository.getOne(uuid, null).getDto();
-    assertEquals(DataExport.ExportStatus.COMPLETED, savedDataExportDto.getStatus());
-
-    // Download and verify CSV has expected structure
-    ResponseEntity<InputStreamResource> response = fileController.downloadFile(uuid,
-        FileController.DownloadType.DATA_EXPORT);
-    assertNotNull(response.getBody());
-
-    try (BufferedReader reader = new BufferedReader(
-        new InputStreamReader(response.getBody().getInputStream(), StandardCharsets.UTF_8))) {
-
-      List<String> lines = new ArrayList<>();
-      String line;
-      while ((line = reader.readLine()) != null) {
-        lines.add(line);
-      }
-
-      // Should have aliased column headers even with minimal data
-      assertTrue(lines.size() >= 2, "Should have header + at least 1 data row");
-      assertTrue(lines.get(0).contains("Project.alias"), "Should have aliased headers");
-      assertTrue(lines.get(1).contains("MINIMAL-SAMPLE-001"), 
-          "Should contain the minimal sample name");
-    }
-
-    // Clean up
-    dataExportRepository.onDelete(uuid);
-    
-    // Clean up test index
-    try {
-      esClient.indices().delete(d -> d.index(testIndex));
-    } catch (Exception e) {
-      // Ignore cleanup errors
-    }
-  }
-
-  /**
-   * Builds a test material sample document with realistic structure matching the provided example.
-   */
-  private String buildTestMaterialSampleDocument(UUID id, String materialSampleName, UUID projectId) {
-    UUID collectionId = UUID.randomUUID();
-    UUID collectingEventId = UUID.randomUUID();
-    
-    return """
-    {
-      "data": {
-        "id": "%s",
-        "type": "material-sample",
-        "attributes": {
-          "version": 1,
-          "group": "test-group",
-          "createdOn": "2025-06-09T16:39:22.29994Z",
-          "createdBy": "test-user",
-          "dwcOtherCatalogNumbers": ["201697"],
-          "materialSampleName": "%s",
-          "materialSampleType": "MIXED_ORGANISMS",
-          "managedAttributes": {
-            "experimental_replicate": "A"
-          },
-          "materialSampleRemarks": "Test sample for ENA export",
-          "allowDuplicateName": false,
-          "isRestricted": false
-        },
-        "relationships": {
-          "collectingEvent": {
-            "data": {
-              "id": "%s",
-              "type": "collecting-event"
-            }
-          },
-          "projects": {
-            "data": [{
-              "id": "%s",
-              "type": "project"
-            }]
-          },
-          "collection": {
-            "data": {
-              "id": "%s",
-              "type": "collection"
-            }
-          }
-        }
-      },
-      "included": [
-        {
-          "id": "%s",
-          "type": "project",
-          "attributes": {
-            "createdOn": "2025-03-27T13:51:41.532518Z",
-            "createdBy": "test-user",
-            "group": "test-group",
-            "name": "GRDI-Eco Invertebrates",
-            "startDate": "2016-04-01",
-            "endDate": "2022-03-31",
-            "status": "Active",
-            "multilingualDescription": {
-              "descriptions": [{
-                "lang": "en",
-                "desc": "GRDI-Ecobiomics test project"
-              }]
-            }
-          }
-        },
-        {
-          "id": "%s",
-          "type": "collection",
-          "attributes": {
-            "createdOn": "2025-03-27T13:14:09.985462Z",
-            "createdBy": "test-user",
-            "group": "test-group",
-            "name": "Test Collection",
-            "code": "TC-001"
-          }
-        },
-        {
-          "id": "%s",
-          "type": "collecting-event",
-          "attributes": {
-            "version": 0,
-            "group": "test-group",
-            "createdBy": "test-user",
-            "createdOn": "2025-06-09T16:39:21.819019Z",
-            "geoReferenceAssertions": [{
-              "dwcDecimalLatitude": 46.45451,
-              "dwcDecimalLongitude": -62.37839,
-              "isPrimary": true
-            }],
-            "dwcVerbatimLocality": "Bear River Road",
-            "dwcRecordedBy": "RS",
-            "startEventDateTime": "2016-10-05",
-            "endEventDateTime": "2016-10-05",
-            "dwcStateProvince": "PEI",
-            "habitat": "riffle",
-            "dwcMinimumElevationInMeters": 11.00,
-            "dwcMinimumDepthInMeters": 21.00
-          }
-        }
-      ]
-    }
-    """.formatted(
-        id.toString(),
-        materialSampleName,
-        collectingEventId.toString(),
-        projectId.toString(),
-        collectionId.toString(),
-        projectId.toString(),
-        collectionId.toString(),
-        collectingEventId.toString()
-    );
-  }
-
-  /**
-   * Builds a minimal material sample document for testing edge cases.
-   */
-  private String buildMinimalMaterialSampleDocument(UUID id, String materialSampleName) {
-    return """
-    {
-      "data": {
-        "id": "%s",
-        "type": "material-sample",
-        "attributes": {
-          "materialSampleName": "%s",
-          "createdBy": "test-user",
-          "createdOn": "2025-06-09T16:39:22.29994Z",
-          "group": "test-group"
-        },
-        "relationships": {}
-      },
-      "included": []
-    }
-    """.formatted(id.toString(), materialSampleName);
-  }
-
-  /**
-   * Helper method to read lines from a ZIP entry without closing the underlying ZipInputStream.
-   */
-  private List<String> readLinesFromZipEntry(ZipInputStream zis) throws IOException {
-    List<String> lines = new ArrayList<>();
-    BufferedReader reader = new BufferedReader(new InputStreamReader(zis, StandardCharsets.UTF_8));
-    String line;
-    while ((line = reader.readLine()) != null) {
-      lines.add(line);
-    }
-    return lines;
-  }
-}
